requirements.txt
streamlit>=1.28.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
plotly>=5.15.0
seaborn>=0.12.0
requests>=2.31.0
networkx>=3.0
folium>=0.14.0
